(Optional)Adding more information to run the project locally
Should we go over how the project should be run locally using the requirements.txt file in more detail? I've added some explanation to the README, but it's entirely optional, and I'm fine if this isn't needed as per the project requirements. That's why I am adding a PR for this